# docs/04-api/Enterprise_AI_API_Gateway_OpenAPI_3.1_V1.2_Final.yaml
openapi: 3.1.0
info:
  title: Enterprise AI & API Gateway
  version: 1.2.0
  summary: Final executable HTTP contract for Enterprise AI & API Gateway V1
  description: Derived from PRD V1.4, HLD V1.1 corrections, ERD V1.2 Final, Physical DB/Alembic V1.2 Final, API Specification
    V1.2.2 Final, frozen M0-M9 execution contracts, and the M2.6 Normal API Registry Contract Reconciliation addendum.
    This artifact consolidates final V1 semantics for Normal API, OpenAI-compatible LLM, Admin browser sessions,
    traffic-admission health, idempotency, certification, credential invalidation, SSRF/TLS and SSE.
servers:
- url: https://gateway.example.com
  description: Production example
security: []
...
paths:
...
  /api/v1/admin/applications:
    get:
      tags:
      - Applications
      operationId: listApplications
      summary: List applications
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Application page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - Applications
      operationId: createApplication
      summary: Create application
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/OptionalIdempotencyKey'
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '409':
          $ref: '#/components/responses/ResourceConflict'
      description: When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.
  /api/v1/admin/applications/{application_id}:
    parameters:
    - $ref: '#/components/parameters/ApplicationId'
    get:
      tags:
      - Applications
      operationId: getApplication
      security:
      - AdminToken: []
      responses:
        '200':
          description: Application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - Applications
      operationId: updateApplication
      security:
      - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationPatch'
      responses:
        '200':
          description: Updated application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '409':
          $ref: '#/components/responses/ResourceConflict'
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      description: Updates application metadata/status. Committed DISABLED status makes all owned gw_ credentials unauthorized
        within <=5 seconds. Session-authenticated mutation requires X-CSRF-Token.
  /api/v1/admin/api-keys:
    get:
      tags:
      - API Keys
      operationId: listApiKeys
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: API key metadata page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyPage'
    post:
      tags:
      - API Keys
      operationId: createApiKey
      summary: Create API key; raw secret is shown once
      description: 'Idempotency-Key is mandatory.


        When authenticated by AdminSessionCookie, X-CSRF-Token is mandatory. Supplying both AdminToken and AdminSessionCookie
        is rejected as ambiguous credentials.'
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/RequiredIdempotencyKey'
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: Raw gw_ key is returned only in this response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        '409': &id001
          description: Idempotency conflict or same operation currently IN_PROGRESS. Inspect normalized error.code. Retry-After
            is present for idempotency_in_progress.
          headers:
            Retry-After:
              description: Present for idempotency_in_progress; V1 value is 1 second. May be omitted for idempotency_conflict.
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/api-keys/{api_key_id}/revoke:
    post:
      tags:
      - API Keys
      operationId: revokeApiKey
      security:
      - AdminToken: []
      parameters:
      - $ref: '#/components/parameters/ApiKeyId'
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Revoked key metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyMetadata'
        '404':
          $ref: '#/components/responses/NotFound'
      description: Committed revoke becomes authorization-effective on the Data Plane within <=5 seconds via targeted invalidation
        plus dedicated credential security-status revalidation. Session-authenticated mutation requires X-CSRF-Token.
...
components:
  securitySchemes:
    GatewayApiKey:
      type: http
      scheme: bearer
      bearerFormat: gw_<secret>
      description: Data Plane credential only.
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: adm_<secret>
      description: Per-admin Control Plane credential only.
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: gw_admin_session
      description: Opaque Redis-backed Admin browser session cookie. Fail-closed when session store is unavailable.
  parameters:
...
    RequiredIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 'Required. Same key + same canonical fingerprint replays the original committed outcome for 24h. Different
        fingerprint => 409 idempotency_conflict. Concurrent IN_PROGRESS => 409 idempotency_in_progress with Retry-After: 1.'
      schema:
        type: string
        minLength: 8
        maxLength: 128
    OptionalIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque server-generated cursor
      schema:
        type: string
    ApplicationId:
      name: application_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ApiKeyId:
      name: api_key_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required only when the operation is authenticated with AdminSessionCookie and the HTTP method is state-changing.
        Not required for direct Bearer adm_ authentication.
      schema:
        type: string
        minLength: 16
        maxLength: 512
...
  responses:
    Unauthorized:
      $ref: '#/components/responses/_Unauthorized'
    _Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found; cross-tenant references also return 404
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ResourceConflict:
      description: Resource/configuration conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      - request_id
      additionalProperties: false
      properties:
        error:
          type: object
          required:
          - message
          - type
          - code
          additionalProperties: false
          properties:
            message:
              type: string
            type:
              type: string
              enum:
              - authentication_error
              - authorization_error
              - rate_limit_exceeded
              - invalid_request
              - resource_not_found
              - model_not_found
              - model_not_allowed
              - provider_timeout
              - provider_rate_limit
              - provider_unavailable
              - provider_authentication_error
              - provider_error
              - all_targets_unavailable
              - gateway_timeout
              - gateway_internal_error
              - resource_conflict
              - route_conflict
              - active_credential_conflict
              - price_window_conflict
              - idempotency_conflict
              - configuration_conflict
              - request_too_large
              - upstream_unavailable
              - upstream_timeout
              - idempotency_in_progress
            param:
              type:
              - string
              - 'null'
            code:
              type: string
            metadata:
              type: object
              additionalProperties: true
              description: Safe aggregate diagnostics only; no provider IDs, base URLs, credentials, or topology.
              properties:
                attempted_targets:
                  type: integer
                  minimum: 0
                ineligible_targets:
                  type: integer
                  minimum: 0
                last_error_code:
                  type:
                  - string
                  - 'null'
        request_id:
          type: string
          format: uuid
          description: Canonical Gateway-generated UUID v7.
...
    Application:
      type: object
      required:
      - id
      - name
      - status
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - DISABLED
    ApplicationCreate:
      type: object
      required:
      - name
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
    ApplicationPatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        name:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - DISABLED
    ApplicationPage:
      type: object
      required:
      - data
      - next_cursor
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Application'
        next_cursor:
          type:
          - string
          - 'null'
    ApiKeyCreate:
      type: object
      required:
      - application_id
      - name
      additionalProperties: false
      properties:
        application_id:
          type: string
          format: uuid
        name:
          type: string
        expires_at:
          type:
          - string
          - 'null'
          format: date-time
    ApiKeyMetadata:
      type: object
      required:
      - id
      - name
      - key_prefix
      - status
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - DISABLED
          - REVOKED
        expires_at:
          type:
          - string
          - 'null'
          format: date-time
    ApiKeyCreated:
      allOf:
      - $ref: '#/components/schemas/ApiKeyMetadata'
      - type: object
        required:
        - key
        properties:
          key:
            type: string
            pattern: ^gw_
            description: Raw secret shown once
    ApiKeyPage:
      type: object
      required:
      - data
      - next_cursor
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyMetadata'
        next_cursor:
          type:
          - string
          - 'null'
...